LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/drawing-rectangles-to-the-canvas/2d.fillRect.clip-expected.txt

2d.fillRect.clip
fillRect is affected by clipping regions
Actual output:
Expected output:


PASS fillRect is affected by clipping regions
